+++ src/odemis/gui/comp/overlay/shapes.py
@@ -23,7 +23,7 @@ from abc import ABCMeta, abstractmethod
 from collections import deque, namedtuple
 from enum import Enum
 import logging
-from typing import Tuple, List, Union, Deque
+from typing import Tuple, List, Union, Deque, Optional
 
 import numpy
 import wx
@@ -75,14 +75,12 @@ class EditableShape(metaclass=ABCMeta):
         return (abs(xmin - xmax), abs(ymin - ymax))
 
     @abstractmethod
-    def is_point_in_shape(self, v_point: Tuple[float, float]) -> bool:
+    def check_point_proximity(self, v_point: Tuple[float, float]) -> bool:
         """
-        Determine if the view point is in the shape.
+        Determine if the view point is in the proximity of the shape.
 
         :param: v_point: The point in view coordinates.
-
-        :returns: whether the view point is inside the shape or not.
-
+        :returns: whether the view point is in proximity of the shape.
         """
         pass
 
@@ -113,6 +111,21 @@ class EditableShape(metaclass=ABCMeta):
     def restore_state(self, state):
         """Restore the shape to a given state."""
         pass
+
+    def on_left_down(self, evt):
+        evt.Skip()
+
+    def on_left_up(self, evt):
+        evt.Skip()
+
+    def on_right_down(self, evt):
+        evt.Skip()
+
+    def on_right_up(self, evt):
+        evt.Skip()
+
+    def on_motion(self, evt):
+        evt.Skip()
 
 
 class ShapesOverlay(WorldOverlay):
@@ -153,13 +166,11 @@ class ShapesOverlay(WorldOverlay):
 
     def clear(self):
         """Remove all shapes and update canvas."""
-        self._activate_shapes(False)
         self._shapes.clear()
 
     def remove_shape(self, shape):
         """Remove the shape and update canvas."""
         if shape in self._shapes:
-            shape.active.value = False
             self._shapes.remove(shape)
             self.cnvs.request_drawing_update()
 
@@ -182,56 +193,51 @@ class ShapesOverlay(WorldOverlay):
         else:
             WorldOverlay.on_leave(self, evt)
 
-    def _activate_shapes(self, flag=True):
-        """Activate or de-activate the shapes."""
-        for shape in self._shapes:
-            shape.active.value = flag
-
-    def _deselected_shapes(self):
+    def _deselect_shapes(self):
         """Deselect shapes except the selected shape."""
         for shape in self._shapes:
             if shape.selected.value and shape != self._selected_shape:
                 shape.selected.value = False
 
-    def _unbind_shape_events(self, shape: EditableShape):
-        """Unbind shape's events bound to the canvas."""
-        shape.cnvs.Unbind(wx.EVT_LEFT_DOWN, handler=shape.on_left_down)
-        shape.cnvs.Unbind(wx.EVT_LEFT_UP, handler=shape.on_left_up)
-        shape.cnvs.Unbind(wx.EVT_RIGHT_DOWN, handler=shape.on_right_down)
-        shape.cnvs.Unbind(wx.EVT_RIGHT_UP, handler=shape.on_right_up)
-        shape.cnvs.Unbind(wx.EVT_MOTION, handler=shape.on_motion)
-
     def _on_tool(self, selected_tool):
         """Update the overlay when it's active and tools change."""
-        if selected_tool == self.tool:
-            self.active.value = True
-            self._activate_shapes(True)
-        else:
-            self.active.value = False
-            self._activate_shapes(False)
-            self.cnvs.reset_default_cursor()
+        self.active.value = selected_tool == self.tool
 
-    def _get_shape(self, v_pos: Tuple[float, float]):
+    def _get_shape(self, v_pos: Tuple[float, float]) -> Optional[EditableShape]:
         """
-        Find a shape corresponding to the given view position.
-        :return: the most appropriate shape.
-            If no shape is found, it returns None.
+        Find the shape corresponding to the given view position based on proximity.
+
+        It tries to find the shape whose center is closest to the view position and then checks if
+        the view position is in the proximity of that shape to confirm it.
+
+        Note:
+        This method considers only 4 shapes with centers closest to the view position. In general
+        the user should always click as close as possbile to the center of the desired shape. If there
+        are overlapping shapes it is advised to click in an area of the desired shape that is not part
+        of the intersection of overlapping shapes. Given the above note, this heuristic might fail
+        in some corner cases.
+
+        :param v_pos: The position in view coordinates.
+        :return: The desired shape, or None if no shape is found.
         """
         if self._shapes:
             v_centers = numpy.array([shape.v_center for shape in self._shapes])
             v_centers_kdtree = cKDTree(v_centers)
-            distances, indices = v_centers_kdtree.query(numpy.array([v_pos[0], v_pos[1]]), k=4)
+             # Query the 4 nearest centers to the given view position
+            distances, indices = v_centers_kdtree.query(numpy.array(v_pos), k=4)
             for distance, index in zip(distances, indices):
-                if not numpy.isinf(distance):
-                    shape = self._shapes[index]
-                    if shape.is_point_in_shape(v_pos):
-                        return shape
+                # Distances are sorted by nearest first
+                # If a distance is infinite, no need to check further as remaining distances will be infinite
+                if numpy.isinf(distance):
+                    return None
+                shape = self._shapes[index]
+                if shape.check_point_proximity(v_pos):
+                    return shape
         return None
 
     def _create_new_shape(self):
         """Create a new shape."""
         shape = self.shape_cls(self.cnvs)
-        self._unbind_shape_events(shape)
         self._shapes.append(shape)
         self.new_shape._set_value(shape, force_write=True)
         return shape
@@ -240,7 +246,6 @@ class ShapesOverlay(WorldOverlay):
         """Copy a selected shape to a view position as the center."""
         # Copy the shape
         shape = self._shape_to_copy.copy()
-        self._unbind_shape_events(shape)
         self._shapes.append(shape)
         self.new_shape._set_value(shape, force_write=True)
         # Move the copied shape to a view position
@@ -265,9 +270,8 @@ class ShapesOverlay(WorldOverlay):
             if self._selected_shape is None:
                 self._selected_shape = self._create_new_shape()
                 self._is_new_shape = True
-            self._selected_shape.active.value = True
             self._selected_shape.on_left_down(evt)
-            self._deselected_shapes()
+            self._deselect_shapes()
         WorldOverlay.on_left_down(self, evt)
 
     def on_char(self, evt):
